modular_django/requirements.txt
asgiref==3.5.2
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
ecdsa==0.18.0
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-core==2.3.2
google-cloud-documentai==2.0.3
google-cloud-storage==2.5.0
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.4
MarkupSafe==2.1.1
martor==1.6.14
msrest==0.7.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.7
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.5.0
pyparsing==3.0.9
pyspellchecker==0.7.0
python-jose==3.3.0
python-keycloak==2.6.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Wand==0.6.10
